+++ tech1-framework-incidents/src/main/java/io/tech1/framework/incidents/domain/IncidentAttributes.java
@@ -9,31 +9,31 @@ public class IncidentAttributes {
     public static class Keys {
         public static final String TYPE = "incidentType";
 
-        public static final String USERNAME = "Username";
-        public static final String PASSWORD = "Password";
+        public static final String USERNAME = "username";
+        public static final String PASSWORD = "password";
 
-        public static final String TRIGGER = "_Trigger";
+        public static final String TRIGGER = "_trigger";
 
-        public static final String EXCEPTION = "Exception";
-        public static final String TRACE = "Trace";
-        public static final String MESSAGE = "Message";
-        public static final String METHOD = "Method";
-        public static final String PARAMS = "Params";
+        public static final String EXCEPTION = "exception";
+        public static final String TRACE = "trace";
+        public static final String MESSAGE = "message";
+        public static final String METHOD = "method";
+        public static final String PARAMS = "params";
 
-        public static final String INVITATION_CODE = "Invitation Code";
+        public static final String INVITATION_CODE = "invitationCode";
 
-        public static final String BROWSER = "Browser";
-        public static final String COUNTRY = "Country";
-        // todo [yy] naming rules
-        public static final String IP_ADDRESS = "IP address";
-        public static final String WHAT = "What";
-        public static final String WHEN = "When";
-        public static final String WHERE = "Where";
+        public static final String BROWSER = "browser";
+        public static final String COUNTRY = "country";
+        public static final String COUNTRY_CODE = "countryCode";
+        public static final String COUNTRY_FLAG = "countryFlag";
+        public static final String IP_ADDRESS = "ipAddress";
+        public static final String WHAT = "what";
+        public static final String WHEN = "when";
+        public static final String WHERE = "where";
     }
 
-    // todo [yy] naming rules
     @UtilityClass
-    public static class Values {
+    public static class IncidentsTypes {
         public static final String THROWABLE = "Throwable";
         public static final String AUTHENTICATION_LOGIN = "Authentication Login";
         public static final String AUTHENTICATION_LOGIN_FAILURE = "Authentication Login Failure";